Add files via upload
This current project was in a development stage and is based on not sensitive data (simulated messages of a Chatbot). 
The pipeline I implemented - not sensitive, as well - is divided into 2 parts; both of them call some Python classes and functions defined in another file which I cannot pubblish because it includes code which Microsoft employees wrote in order to perform the correct setup on Databricks.